Ghost
Play Ghost:

A party of people ranging from 3-100 can come together and play a game of Ghost using this app. Just input the number for each type of character and share the game code! Once everyone inputs the game code on the home page. They are in the game and will get assigned their distinguishable character once everyone has joined.