Couch Game
Play Couch Game:

A party of people ranging from 8-100 can come together and play a game of Couch Game using this app. Just input team names and the number of people playing (which should be even) and share the game code! Once everyone inputs the game code on the home page, they are in the game and will input their names (and choose their teams if the teams are not random) and submit. Then, they will get assigned another person's name within the group once everyone has joined.